restructure files for incoming linter addition
