+++ src/parser/index.ts
@@ -1,5 +1,5 @@
-import { parse as peggyParse } from "./peggy-parser/generated";
-import { ER, ERSchema } from "./types/parser/ER";
+import { parse as peggyParse } from "./peggy/generated";
+import { ER, ERSchema } from "../types/parser/ER";
 
 // type safe wrapper for the parser generated by peggy
 /**
